reimplemented OTU class. some major refactoring of pylab.base into a submodule.
