+++ mdsine2/__init__.py
@@ -34,7 +34,7 @@ from .pylab.multiprocessing import ispersistentworker, ispersistentpool, isDASW,
 
 # Import commonly used Pylab functions and classes
 from .pylab.util import toarray, fast_index, coarsen_phylogenetic_tree
-from .pylab.base import Taxon, OTU, TaxaSet, qPCRdata, Saveable, Traceable, BasePerturbation, \
+from .pylab.base import Taxon, OTU, TaxaSet, qPCRdata, Saveable, TraceableNode, BasePerturbation, \
     Subject, Study, condense_matrix_with_taxonomy, taxaname_for_paper, TAXANAME_PAPER_FORMAT, \
     taxaname_formatter, Perturbations, qPCRdata
 from .pylab.variables import Variable, Constant, summary
